+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Support for inline badge block `google-customer-review-badge`.
+
 ## [1.0.7] - 2020-11-10
 
 ### Fixed